# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Finance dashboard API
  description: |
    Main routes only. Behaviour, auth, and examples: see README and API Testing.md.
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://{host}
    description: Deployed (set host)
    variables:
      host:
        default: api.example.com
...
paths:
...
  /api/finance/records:
    get:
      tags: [Finance]
      summary: List records (analyst or admin)
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100 }
        - name: type
          in: query
          schema: { type: string, enum: [income, expense] }
        - name: category
          in: query
          schema: { type: string }
        - name: dateFrom
          in: query
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Paginated list
        "403":
          description: Insufficient role
    post:
      tags: [Finance]
      summary: Create record (admin only)
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinanceRecordBody"
      responses:
        "201":
          description: Created
        "400":
          description: Validation error
        "403":
          description: Not admin
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  schemas:
...
    FinanceRecordBody:
      type: object
      required: [amount, type, category, date]
      properties:
        amount:
          type: number
          minimum: 0
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        date:
          type: string
          format: date
        notes:
          type: string